+++ libeasea/include/CIndividual.h
@@ -22,14 +22,14 @@ public:
 	virtual ~CIndividual();
 	virtual float evaluate() = 0;
 	virtual void printOn(std::ostream& O) const = 0;
-	virtual size_t mutate(float pMutationPerGene) = 0;
+	virtual unsigned mutate(float pMutationPerGene) = 0;
 	virtual CIndividual* crossover(CIndividual** p2) = 0;
 	virtual CIndividual* clone() = 0;
 
         virtual std::string serialize() = 0;
         virtual void deserialize(std::string EASEA_Line) = 0;
 
-	static size_t getCrossoverArrity(){ return 2; }
+	static unsigned getCrossoverArrity(){ return 2; }
 	float getFitness(){ return this->fitness; }
